Task 26. Independent class "student".
